# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: The authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /emails:
    get:
      operationId: listEmails
      tags: [emails]
      summary: List emails
      description: List the authenticated user's Gmail messages with optional filtering by label, view, and search. Returns one bounded page plus a cursor for the next.
      parameters:
        - name: labelId
          in: query
          required: false
          schema:
            type: string
          description: Filter to emails that have this Gmail label id
        - name: view
          in: query
          required: false
          schema:
            type: string
            enum: [all, unlabeled, starred, unread]
          description: Predefined filter view
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Gmail search query across subject, sender and body
        - name: pageToken
          in: query
          required: false
          schema:
            type: string
          description: Opaque cursor from a previous response's nextPageToken to fetch the next page
      responses:
        "200":
          description: A page of emails
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailPage"
  /emails/{id}:
    get:
      operationId: getEmail
      tags: [emails]
      summary: Get a single email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateEmail
      tags: [emails]
      summary: Update an email's read/starred state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailUpdate"
      responses:
        "200":
          description: Updated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
  /emails/{id}/labels:
    put:
      operationId: setEmailLabels
      tags: [emails]
      summary: Replace the set of user labels on an email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailLabelInput"
      responses:
        "200":
          description: Updated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
  /emails/{id}/labels/{labelId}:
    delete:
      operationId: removeEmailLabel
      tags: [emails]
      summary: Remove a single label from an email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: labelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
  /emails/bulk-label:
    post:
      operationId: bulkLabelEmails
      tags: [emails]
      summary: Apply or remove labels across multiple emails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkLabelInput"
      responses:
        "200":
          description: The affected emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Email"
  /emails/bulk-action:
    post:
      operationId: bulkEmailAction
      tags: [emails]
      summary: Perform a mailbox action across multiple emails
      description: Archive, trash, mark spam, or change read/starred state for many emails at once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkActionInput"
      responses:
        "200":
          description: The affected emails (fresh state)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Email"
  /emails/send:
    post:
      operationId: sendEmail
      tags: [emails]
      summary: Send, reply to, or forward an email
      description: Composes and sends a plain-text message via the Gmail send API. Supplying inReplyToId threads the message as a reply.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmailInput"
      responses:
        "200":
          description: The sent message identifiers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendEmailResult"
  /emails/{id}/unsubscribe:
    post:
      operationId: unsubscribeEmail
      tags: [emails]
      summary: One-click unsubscribe via the List-Unsubscribe header
      description: Performs an RFC 8058 one-click unsubscribe POST when supported, otherwise returns a link for the client to open.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The unsubscribe outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsubscribeResult"
  /labels:
    get:
      operationId: listLabels
      tags: [labels]
      summary: List labels with email counts
      responses:
        "200":
          description: A list of labels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Label"
    post:
      operationId: createLabel
      tags: [labels]
      summary: Create a label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelInput"
      responses:
        "201":
          description: Created label
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Label"
  /labels/{id}:
    patch:
      operationId: updateLabel
      tags: [labels]
      summary: Update a label
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelUpdate"
      responses:
        "200":
          description: Updated label
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Label"
    delete:
      operationId: deleteLabel
      tags: [labels]
      summary: Delete a label
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        picture:
          type: ["string", "null"]
        autoLabelEnabled:
          type: boolean
          description: Whether background auto-labeling is enabled for this user.
        dailyDigestEnabled:
          type: boolean
          description: Whether daily unread label digests are emailed to the user.
...
    Label:
      type: object
      required: [id, name, color, isSystem, emailCount]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: ["string", "null"]
          description: A hex color string, e.g. "#6366f1"
        description:
          type: ["string", "null"]
        isSystem:
          type: boolean
        emailCount:
          type: integer
    LabelInput:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        description:
          type: string
    LabelUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        description:
          type: ["string", "null"]
    Email:
      type: object
      required:
        - id
        - sender
        - senderEmail
        - subject
        - snippet
        - body
        - bodyHtml
        - receivedAt
        - isRead
        - isStarred
        - labels
        - hasRemoteImages
        - unsubscribeUrl
        - unsubscribeMailto
      properties:
        id:
          type: string
        sender:
          type: string
        senderEmail:
          type: string
        subject:
          type: string
        snippet:
          type: string
        body:
          type: string
          description: Plain-text rendering of the email body (HTML stripped). Used for previews and AI context.
        bodyHtml:
          type: string
          description: Sanitized HTML body for rich display. Empty string when the email has no HTML part. Remote image sources are held in data-blocked-src until the user displays images.
        receivedAt:
          type: string
          format: date-time
        isRead:
          type: boolean
        isStarred:
          type: boolean
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
        hasRemoteImages:
          type: boolean
          description: True when the sanitized HTML has remote images that were held back (blocked by default).
        unsubscribeUrl:
          type: ["string", "null"]
          description: One-click/managed https unsubscribe target from List-Unsubscribe, if any.
        unsubscribeMailto:
          type: ["string", "null"]
          description: A mailto unsubscribe target from List-Unsubscribe, if any.
    EmailPage:
      type: object
      required: [emails, nextPageToken]
      properties:
        emails:
          type: array
          items:
            $ref: "#/components/schemas/Email"
        nextPageToken:
          type: ["string", "null"]
          description: Cursor to fetch the next page; null when the window is exhausted.
    EmailUpdate:
      type: object
      properties:
        isRead:
          type: boolean
        isStarred:
          type: boolean
    EmailLabelInput:
      type: object
      required: [labelIds]
      properties:
        labelIds:
          type: array
          items:
            type: string
    BulkLabelInput:
      type: object
      required: [emailIds, labelId, action]
      properties:
        emailIds:
          type: array
          items:
            type: string
        labelId:
          type: string
        action:
          type: string
          enum: [add, remove]
    BulkActionInput:
      type: object
      required: [emailIds, action]
      properties:
        emailIds:
          type: array
          items:
            type: string
        action:
          type: string
          enum: [archive, trash, untrash, spam, markRead, markUnread, star, unstar]
    SendEmailInput:
      type: object
      required: [to, subject, body]
      properties:
        to:
          type: string
          description: Recipient address(es), comma-separated.
        cc:
          type: string
        subject:
          type: string
        body:
          type: string
          description: Plain-text message body.
        inReplyToId:
          type: string
          description: Source message id when replying/forwarding; threads the message.
    SendEmailResult:
      type: object
      required: [id, threadId]
      properties:
        id:
          type: string
        threadId:
          type: ["string", "null"]
    UnsubscribeResult:
      type: object
      required: [status, url]
      properties:
        status:
          type: string
          enum: [posted, open, none]
          description: posted = one-click POST done; open = client should open url; none = no target.
        url:
          type: ["string", "null"]
...