# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  picture?: string | null;
  /** Whether background auto-labeling is enabled for this user. */
  autoLabelEnabled?: boolean;
  /** Whether daily unread label digests are emailed to the user. */
  dailyDigestEnabled?: boolean;
}